Updated catch errors to display msg

+++ back-end/controllers/stuff.js
@@ -64,19 +64,34 @@ exports.createThing = ('/', (req, res, next) =>{
   };
 
   exports.deleteThing = (req, res, next) => {
-    Thing.deleteOne({_id: req.params.id}).then(
-      () => {
-        res.status(200).json({
-          message: 'Deleted!'
-        });
+    Thing.findOne({_id: req.params.id}).then (
+      (thing) => {
+        if (!thing) {
+          return res.status(404).json({
+            error: new error('No such thing!')
+          });
+        }
+        if (thing.userId !== req.auth.userId) {
+          return res.status(400).json({
+            error: new error('Unauthorized request!')
+          });
+        }
+        Thing.deleteOne({_id: req.params.id}).then(
+          () => {
+            res.status(200).json({
+              message: 'Deleted!'
+            });
+          }
+        ).catch(
+          (error) => {
+            res.status(400).json({
+              error: error
+            });
+          }
+        );
       }
-    ).catch(
-      (error) => {
-        res.status(400).json({
-          error: error
-        });
-      }
-    );
+    )
+   
   };
 
   exports.getAllThings = (req, res, next) => {
